requirements.txt
moviepy==1.0.3
instaloader
requests
openai-whisper